get all order updated by added search with email

+++ src/modules/orders/order.validation.ts
@@ -4,10 +4,7 @@ const orderValidation = z.object({
   productId: z
     .string()
     .length(24, { message: "Product ID must be exactly 24 characters long" }),
-  price: z
-    .number()
-    .int()
-    .positive({ message: "Price must be a positive integer" }),
+  price: z.number().positive({ message: "Price must be a positive integer" }),
   quantity: z
     .number()
     .int()
